// src/app/models/admin/dangkyhoatdongkhoangsan/dangkykhaithac/dkthamdochuyennhuong.model.ts
export class OutputDkThamDoChuyenNhuongModel {
  iddangkythamdo: string;
  idhoso: string;
  idgiayphep: string;
  sohopdong: string;
  siteid: string;
  sogiayphep: string;
  ngayhopdong: string;
  loaidoituong: string;
  idcanhantochuc: string;
  tencanhantochuc: string;
  sogiayto: string;
  loaigiayto: number;
  ngaycap: string;
  noicap: string;
  diachi: string;
  dienthoai: string;
  fax: string;
  email: string;
  website: string;
}

export class InputDkThamDoChuyenNhuongModel {
  iddangkythamdo: string;
  idhoso: string;
  idgiayphep: string;
  siteid: string;
  ngayhopdong: string;
  sohopdong: string;
}